feature: save Auth information into indexDB
