+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@notionhq/client": "^2.2.3",
     "cheerio": "1.0.0-rc.12",
+    "idb-keyval": "^6.2.0",
     "vue": "^3.2.45",
     "vue-router": "4"
   },
