+++ src/types.ts
@@ -47,12 +47,16 @@ export interface OneTimeMessage {
 	/* sender: popup script
 	 * receiver: background script
 	 * notify background script check the token secret and database id provided by user */
-	checkAuth?: { tokenSecret: string; databaseID: string };
+	checkAndSaveAuth?: { tokenSecret: string; databaseID: string };
 }
 
 export type ExtensionError =
 	| {
 			name: "ContentScriptError";
 			message: "Sorry, D2N was unable to retrieve information about the book from this page. Please make sure you are on a valid book page and try again";
+	  }
+	| {
+		name: "Token&IDNotExistError";
+			message: "Sorry, You have to provide Notion Secret Token and Database ID to use D2N"
 	  }
 	| NotionClientError;
